requirements.txt
pillow==9.2.0
rembg